tifl_project/requirements.txt
tensorflow==2.12.0
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
pyyaml==6.0
tqdm==4.65.0
